Continued  work on poc with state machine
* added state machine of stargate project, so we can define the different states easier with guard functions and everything
* added first simple recipe, so we should be able to test an easier use case

+++ OctoAwesome/OctoAwesome.PoC/RecipeItem.cs
@@ -1,3 +1,3 @@
 ﻿namespace OctoAwesome.PoC;
 
-public readonly record struct RecipeItem(string ItemName, decimal Count, string? MaterialName = null);
+public readonly record struct RecipeItem(string ItemName, int Count, string? MaterialName = null);
